[Refactor] 생성 이벤트 패키지로 분리
