+++ src/main/kotlin/com/whatever/caro/card/internal/deck/event/created/DeckCreatedEvent.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.card.internal.deck.event
+package com.whatever.caro.card.internal.deck.event.created
 
 data class DeckCreatedEvent(
     val deckId: Long,
